+++ derive_builder/src/options/struct_options.rs
@@ -1,5 +1,5 @@
 use syn;
-use derive_builder_core::{DeprecationNotes, BuilderPattern, Builder, BuildMethod};
+use derive_builder_core::{DeprecationNotes, BuilderPattern, Builder, BuildMethod, Bindings};
 use options::DefaultExpression;
 
 /// These struct options define how the builder is generated.
@@ -22,8 +22,8 @@ pub struct StructOptions {
     pub deprecation_notes: DeprecationNotes,
     /// Number of fields on the target struct.
     pub struct_size_hint: usize,
-    /// Whether the generated code should comply with `#![no_std]`.
-    pub no_std: bool,
+    /// Bindings to libstd or libcore.
+    pub bindings: Bindings,
     /// Default expression for the whole struct, e.g. `#[builder(default)]` (default to None).
     pub default_expression: Option<DefaultExpression>,
 }
@@ -54,8 +54,10 @@ impl StructOptions {
             target_ty_generics: Some(ty_generics),
             initializers: Vec::with_capacity(self.struct_size_hint),
             doc_comment: None,
-            no_std: self.no_std,
-            default_struct: self.default_expression.as_ref().map(|x| { x.parse_block() }),
+            bindings: self.bindings,
+            default_struct: self.default_expression
+                .as_ref()
+                .map(|x| { x.parse_block(self.bindings.no_std) }),
         }
     }
 }