+++ derive_builder_core/CHANGELOG.md
@@ -2,12 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## Unreleased
+### Added
+- Bindings to abstract over libstd/libcore
+
+### Changed
+- Use `bindings: Bindings` instead of `no_std: bool`
 
 ### Fixed
 - support generic references in structs #55
+- no_std support #63
 
 ## 0.1.1 - 2017-04-08
+### Added
 - struct default
 
 ## 0.1 - 2017-03-25